refactor: restructure folder structure to make sure each constant and interface is in a file. Also, upadated dependencies.
