+++ .changeset/stupid-onions-change.md
@@ -1,0 +1,5 @@
+---
+'@watergis/maplibre-gl-terradraw': patch
+---
+
+refactor: restructure folder structure to make sure each constant and interface is in a file. Also, upadated dependencies.
