+++ package.json
@@ -44,11 +44,11 @@
 	"devDependencies": {
 		"@changesets/cli": "^2.27.10",
 		"@floating-ui/dom": "^1.6.12",
-		"@playwright/test": "^1.49.0",
+		"@playwright/test": "^1.49.1",
 		"@skeletonlabs/skeleton": "^2.10.3",
 		"@skeletonlabs/tw-plugin": "^0.4.0",
 		"@sveltejs/adapter-auto": "^3.3.1",
-		"@sveltejs/kit": "^2.9.0",
+		"@sveltejs/kit": "^2.10.1",
 		"@sveltejs/package": "^2.3.7",
 		"@sveltejs/vite-plugin-svelte": "^5.0.1",
 		"@tailwindcss/forms": "^0.5.9",
@@ -57,7 +57,7 @@
 		"@turf/centroid": "^7.1.0",
 		"@turf/distance": "^7.1.0",
 		"@types/eslint": "^9.6.1",
-		"@types/node": "^22.10.1",
+		"@types/node": "^22.10.2",
 		"autoprefixer": "^10.4.20",
 		"dependency-cruiser": "^16.7.0",
 		"eslint": "^9.16.0",
@@ -73,13 +73,13 @@
 		"publint": "^0.2.12",
 		"sass": "^1.82.0",
 		"sass-embedded": "^1.82.0",
-		"svelte": "^5.9.0",
+		"svelte": "^5.11.2",
 		"svelte-check": "^4.1.1",
 		"tailwindcss": "^3.4.16",
 		"terra-draw": "1.0.0-beta.8",
 		"typedoc": "^0.27.4",
 		"typescript": "^5.7.2",
-		"typescript-eslint": "^8.17.0",
+		"typescript-eslint": "^8.18.0",
 		"vite": "^6.0.3"
 	},
 	"peerDependencies": {
