+++ src/lib/constants/index.ts
@@ -1,4 +1,5 @@
 export * from './AvailableModes.js';
+export * from './AvailableMeasureModes.js';
 export * from './defaultControlOptions.js';
 export * from './defaultMeasureControlOptions.js';
 export * from './getDefaultModeOptions.js';
