+++ src/lib/controls/MaplibreTerradrawControl.ts
@@ -10,8 +10,8 @@ import type {
 	EventType,
 	TerradrawMode,
 	TerradrawModeClass
-} from './interfaces/index.js';
-import { defaultControlOptions, getDefaultModeOptions } from './constants/index.js';
+} from '../interfaces/index.js';
+import { defaultControlOptions, getDefaultModeOptions } from '../constants/index.js';
 
 /**
  * Maplibre GL Terra Draw Control
