+++ src/lib/index.ts
@@ -1,7 +1,3 @@
-import { MaplibreTerradrawControl } from './MaplibreTerradrawControl.js';
-import { MaplibreMeasureControl } from './MaplibreMeasureControl.js';
-
-export { MaplibreTerradrawControl, MaplibreMeasureControl };
-
+export * from './controls/index.js';
 export * from './constants/index.js';
 export * from './interfaces/index.js';
