+++ src/lib/interfaces/MeasureControlMode.ts
@@ -1,0 +1,6 @@
+import type { AvailableMeasureModes } from '../constants/AvailableMeasureModes.js';
+
+/**
+ * Terra Draw Measure control mode
+ */
+export type MeasureControlMode = (typeof AvailableMeasureModes)[number];
