+++ src/lib/interfaces/TerradrawMode.ts
@@ -1,11 +1,6 @@
-import { AvailableModes, AvailableMeasureModes } from '../constants/AvailableModes.js';
+import { AvailableModes } from '../constants/AvailableModes.js';
 
 /**
  * Terra Draw drawing mode
  */
 export type TerradrawMode = (typeof AvailableModes)[number];
-
-/**
- * Terra Draw Measure control mode
- */
-export type MeasureControlMode = (typeof AvailableMeasureModes)[number];
