+++ src/lib/interfaces/index.ts
@@ -1,5 +1,6 @@
 export * from './TerradrawControlOptions.js';
 export * from './EventTypes.js';
+export * from './MeasureControlMode.js';
 export * from './MeasureControlOptions.js';
 export * from './ModeOptions.js';
 export * from './TerradrawMode.js';
